fixed errors in getCountValidPredictionsOver10ExclCurrent and getDbUserPickValue methods

+++ src/test/java/com/sapfir/helpers/PredictionOperations.java
@@ -77,12 +77,12 @@ public class PredictionOperations {
         return dbOp.getSingleValue(conn, "market", sql);
     }
 
-    public int getDbUserPickValue(String predictionId) {
+    public float getDbUserPickValue(String predictionId) {
         String sql = "select user_pick_value from prediction where id = '" + predictionId + "';";
 
         DatabaseOperations dbOp = new DatabaseOperations();
         String stringUserPickValue = dbOp.getSingleValue(conn, "user_pick_value", sql);
-        return Integer.parseInt(stringUserPickValue);
+        return Float.parseFloat(stringUserPickValue);
     }
 
     private boolean resultDifferent(String predictionID) {
